Changed ADbasic code again
For the DIGPROG1 and DIGPROG2, I had set bit 0 and bit 7 to 1, instead of bit 8 and bit 0, so half of the channels were being considered as inputs. Instead of 0FFh (which does sets bits 0-7 as 1) I am using 0FFFh which will set all bits from 0-11 as 1. Only bit 8 and bit 0 are used, the rest are ignored.