Deleting comments on running iterations
To my sadness, deleting the run list buttons that is reminiscent of cicero.  This is because it would require changign signficant aspects of how we "scan parameters" in the code. I figure it's better to make a GUi that adapted to our code than some "ideal" i have from experience (which isnt' necessarily good)